Removing libconninet 3rdparty component.
The use of libconninet caused Qt to have a cyclic build dependency.
Which was solved by embedding a copy as a 3rd party library. Maemo has
since donated the libconninet code to Qt because (1) it was a private
Maemo API and (2) the Qt ICD plugin is the only user of the library.

This commit moves the relevent code from src/3rdparty/libconninet to
src/plugins/bearer/icd and deletes the rest.

Task-number: QT-3893