Introduction of new memory management. No raw pointers anymore!
Two new classes were introduced; QScriptSharedDataPointer and
QScriptPassPointer. The first is used for normal reference counting, it
is almost the same as the QExplicitlySharedDataPointer. The second is
used only for returning data from a function. It assure that the data
will be destroyed correctly (for example by a QScriptSharedDataPointer
instance).