Fixed regression when typing in QTextControl based widgets on Symbian
The bug was that when querying for the maximum text length, the case
where an invalid QVariant was returned (which is allowed) was not
handled properly. This would lead to input being blocked by the input
context when it shouldn't.

RevBy:    Sami Merila
(cherry picked from commit 3bb3af84bef3c0472ca8ed0d5c6bb3c82320956d)