Fixed performance issue in QML clipping with OpenGL 2.0 paint engine.
Change 4c515ceb fixed Intersect and Unite-clipping after doing
setClipping(false), but also introduced a performance problem.

If there was no pre-existing clip we'd convert IntersectClips to
ReplaceClips, which is unnecessary since the paint engines already
handle this correctly.

IntersectClips can be handled much more efficiently in the OpenGL 2
paint engine for example. We only need to convert to ReplaceClip when
someone has used setClipEnabled(false) which is anyways expensive and
not recommended.

Reviewed-by: Trond