Fix a color propagation issue with Combobox line edit
Some pixmap based themes were affected by this issue since they
needed to resolve the palette against the widget and the lineedit
palette was already resolved. We now set the palette on the lineedit
when you modify the parent palette. You can still maintain a separate
palette on the lineedit, but you have to set it after the parent has
been set. We can remove some workarounds for plastique and vista style
due to this.

Task-number: QTBUG-5950
Reviewed-by: ogoffart