Implemeting, exporting and autotesting QFont::lastResortFont()
An implementation of QFont::lastResortFont() is still(!) missing
in Qt 4.7.0. I only became aware of QTBUG-6921, lately.

This patch...
1) implements QFont::lastResortFont() in qfont_s60.cpp by first
trying to get the lastResortFamily() and then falling back to a
hardcoded font.

2) updates the .def files with one additional entry

3) adds an autotest which verifies that lastResortFamily() does
return a non-empty string. In the firt place, that autotest makes
sure that lastResortFamily() is implemented and exported, so that
something like this issue will not go unnoticed in the next
Qt port.

Task-number: QTBUG-6921
Reviewed-by: Eskil