Fixed antialiased rasterization bug in raster engine.
When rasterization in the gray raster fails due to out of memory there
might already have been a number of spans flushed. To avoid flushing
these spans multiple times and thus getting overdraw artifacts we need
to keep track of how many spans to skip when we redo the rasterization.

This fixes the rendering error in arthur test paths_aa.qps

Reviewed-by: Yoann Lopes