Fix QFontMetrics::lineWidth() for fonts with defined point size
QFontMetrics::lineWidth() and ::underlinePos() return value 1
regardless of the font size if the size was defined in points
(instead of pixels).

(On Symbian) QFontMetrics::lineWidth() calls
QFontEngine::lineThickness() which uses its fontDef.pixelSize in
order to come up with a suitable line width. If the QFont size
was defined in points, Qt needs to make sure that fontDef.pixelSize
is set accordingly. This patch adds the code to make sure that
QFontEngine::fontDef always has a valid pixel size.

tst_QFontMetrics::lineWidth() was added, wich failed before and
passes after this patch.

Task-Number: QTBUG-13009
Autotest: Passes
Reviewed-By: Eskil