Make texture glyph cache on OpenGL work without a paint engine
When used from outside of Qt, the OpenGL texture glyph cache would fall
back to using an image backend. This was not exactly ideal, so instead
it's rewritten to set up its own blit program etc. when the paint engine
is not available. Note that fillInPendingGlyphs() will potentially
alter the current GL state/viewport/bindings/etc., so it should be
called before setting up the actual painting.

Reviewed-by: Kim