Added support for updating the baselines via the generated HTML reports.
The baselineserver can now run as a CGI binary if it detects that
the QUERY_STRING variable is set. Currently it only handles
commands for updating the baselines for failed tests.
For this to work properly the baselineserver binary needs to have
the setuid bit set, so that when it's run as a CGI binary, it will have
the proper rights to move files around. The server currently
assumes the CGI binary is a link to the baselineserver executable
named 'server.cgi' in the configured 'cgi-bin' directory.