Use setParentItem() instead of setParentItemHelper if componentComplete is true
This ensures that if the component has already been completed, itemChange()
is called. This is required to modify QDeclarativePositions in the visual
editor, while keeping a legal state. Without this patch notifications are
missing and the Positioner keeps track of already deleted children.
This resulted in a crash.

Reviewed-by: Thomas Hartmann
(cherry picked from commit 1f8bb573f06234a3d13fb57de5eb644824d5024f)