Fix a custom global object.
Nonexistent properties of a custom global object should be marked as
invalid, not undefined. Please notice that in the QtScript undefined
values are valid.

The patch introduces a new definition of QSVP::isValid; a invalid value
may be bound to an engine, which means that checks for isJSBased() or
engine() are not sufficient to say that value is valid or not anymore.