Track average wait times under our maximum spin time threshold
Further observation shows that spin times slightly over the average wait time
produce the best results. This change keeps a heavily weighted average
of the wait times under 1.5ms (1.5 times the max spin threshold), and adjusts
the spin time to be 150% of the average wait time.

Introduce spin time adjustments when spin locking works, and adjust to
between 150% of the average wait time and the maximum threshold.

Reviewed-by: joao