Extended the convenience functions for QInputDialog for IM hints.
Input method hints make a lot of sense when requesting specific
information from the user, such as for example a URL. Therefore we
include input method hints in the static convenience function.

To maintain BC we need to keep the old symbol around, so a new
function was added. To maintain SC, there cannot be any ambiguities
when deciding which function to link to, therefore the function with
the fewest arguments get to keep all the default arguments, whereas
the new function gets a new explicit argument.

AutoTest: Passed
Task:     QTBUG-13200
RevBy:    Denis Dzyubenko