Create QScopedResource class
This has similar purpose to QScopedPointer, and is implemented using it.
Symbian has many resource classes, e.g. RFile, RSocket, which behave as
value types, but require the Close() member function to be called to release
the server side resource owned by the handle.
This class can be used to give exception safety, for example when trying to
add a resource class to a container std::bad_alloc could be thrown.
By assigning the resource to QScopedResource when it is opened, and calling
take() after it is added to the container, it is ensured that the resource
will be closed if an exception is thrown.

Equivalent to the symbian pattern:
   CleanupClosePushL(file);
   container.AppendL(file); //transfers ownership, may leave
   CleanupStack::Pop(file);

With STL exceptions we can do:
   QScopedResource ptr(file);
   container.append(file); //transfers ownership, may throw
   ptr.take;

Reviewed-By: mread