Added qmake check for presence of RHttpDownloadMgr header
downloadmgrclient.h is not found on S^4 baselines, causing a build
failure.  This commit is a temporary workaround, which disables
progressive download support if the header is not found.

The correct solution is to determine whether the RHttpDownloadMgr
definition has moved, and if so, to modify the .pro file to
include the new path.

Task-number: QTBUG-10769
Reviewed-by: TrustMe