Ensure that the underline is only drawn when expected for an accel
On Mac, the underline for an accel should not be drawn at all as this is
not part of the style guidelines.  This ensures that it is not drawn.

Autotest included which required a change to drawItemText() to go via
the proxy as well so that the autotest would be useful too.

Task-number: QTBUG-14172
Reviewed-by: Richard