make custom command handling in vc(x)proj files sane
instead of glueing the commands with "&&" and having a monster hack
(\r\h) for the case where they cannot be and-ed, glue them with newlines
and error checks. this ensures that we a) never have excessively long
lines, b) don't need the monster hack and c) commands always have proper
error checks, just like in makefiles.

Reviewed-by: mariusSO