Fix copying large data to non-Qt applications
In send_selection when we decide to use INCR properties we set the
given property value to be INCR atom, however we were also sending
SelectionNotify event containing the name of the property that the
client should watch as INCR, which is wrong, we should send the
original property that we were asked to use instead.

This fixes copying large junks of data from Qt application, for
example copying text (>200k) to gedit or images to gimp.

Task-number: QTBUG-8644
Reviewed-by: Bradley T. Hughes