qmake: Ensure right library order when reducing duplicate libraries
Commit 3ed9e2788e8 changed the logic from using lookahead to checking
if the library already existed in the arguments parsed so far. This
broke the case -lFoo -lBar -lFoo where Bar depends on symbols in Foo.
Since we were reducing this to -lFoo -lBar, the linker would not know
which missing symbols to include from Foo when encountering the -lFoo
argument.

We now keep the library order when building the final argument list,
by making sure the position of a duplicated library is always that
of the last instance of that library.

Task-number: QTBUG-13944
Reviewed-by: Simon Hausmann <simon.hausmann@nokia.com>