Fix crash when constructing QThreadStorage after global destructors have run
In this particular case, the destructors vector used by
QThreadStorageData has already been destroyed, yet a new QThreadStorage
is being allocated. This can only happen during global destruction, at
which point we assume that there is only one thread. In order to keep
QThreadStorage working, we need somewhere to store the data, and the
best place we have in this situation is at the tail of the current
thread's tls vector. The destructor is ignored, since we have no where
to store it, and no way to actually call it.

Task-number: QTBUG-10084
Reviewed-by: thiago
Reviewed-by: olivier