Fixed a buffer overrun when pasting large data from non-Qt apps
When we receive a COMPOUND_TEXT data and convert it to multi-byte, if
the new length is less than the original buffer, we resize the buffer,
but then we need to make sure we reset the buffer_offset variable,
which is returned as a "size" of the buffer to a caller.

The conversion to COMPOUND_TEXT format was totally wrong, we were using
the wrong length of the input text, so the converted string was
truncated.

Task-number: QTBUG-8644
Reviewed-by: Bradley T. Hughes
(cherry picked from commit 4b81cb847647450f4bad8a0d9a278d43ebdfecc6)