Refactored PaletteEditorAdvanced in qtconfig
The widget was redrawn and now makes use of QDialogButtonBox. The
ColorButton's are created more correctly for now. Also, all of the Qt4
ColorRoles are now supported.

All the legacy code was dropped so it must not depend of qt3support.
Some refactoring job also was done and much of unused code was removed.

Merge-request: 2286
Reviewed-by: Benjamin Poulain <benjamin.poulain@nokia.com>