Fix text rendering in GL when using the broken-fbo-fallback
This is a backport of 3874cd95e203da40d5205ef6455d7f56cba6923a which
was committed to qt:master. Since the patch that originally exposed
the bug was added to 4.7.2, this patch is required there as well.
Bug is visible when resizing the glyph cache on devices where
QGLContext::brokenFBOReadback defaults to true, or by setting this to
true on desktop.

Reviewed-by: Gunnar