Implement support for DEPLOYMENT.display_name in Symbian
The default package name and the default name that appears in
the application menu is derived from the TARGET variable.
Often the default is not optimal for displaying to end user.
To set a better display name for these purposes DEPLOYMENT.display_name
variable can now be used.

Task-number: QTBUG-14280
Reviewed-by: Janne Anttila