Fix crash when using Q_GLOBAL_STATIC(QWidget...)
If Q_GLOBAL_STATIC is used with a QWidget (or subclass) then the
destructor of QWidget will be executed after the destructor of
QApplication. Since ~QApplication() destroys the S60 environment and
the trap handler, we need to be sure that if QApplication is destroyed,
we do not attempt to use anything from the S60 environment. This
includes RWsSession and the trap handler.

The fix is to avoid flushing the WSERV buffer if QApplication has been
deleted already.

Reviewed-by: axis