Fix memory leak in QPixmap::toSymbianRSgImage() when an error occurs.
In the cases where an error occured while converting a QPixmap to a
VGImage this function would return without deleting the RSgImage
pointer that it created. Fix is to use a QScopedPointer instead. Also
don't use q_check_ptr() since this isn't a CBase derived class.

In case you are wondering why I didn't use a custom deleter here so
that Close() was also called, we need to make sure that Close() is
called on the RSgImage instance before calling Close() on the driver.

Reviewed-by: mread