Fix infinite loop when justifying undisplayable Arabic text
If the Arabic text is for some reason undisplayable, e.g. because of
QTBUG-13132, the font engine will be unable to find the tatweel character
and the kashida width may be returned as 0. This would potentially cause
an infinite loop, as "need" would remain >= minKashida forever because
x - 0 is still >= 0.

Task-number: QTBUG-13130
Reviewed-by: Lars