Crash fix when using the runtime graphics system on Symbian.
If QWidget::destroy() is called (perhaps by setParent) then the window
surface is deleted by the destructor of QSymbianControl instead of in
the "normal" place which is ~QWidget(). This means that we should not
attempt to use the RWindow in the window surface because it is in the
process of being destroyed.

Reviewed-by: Jani Hautakangas