Create QSystemError class
This class is for holding errors returned from system calls, which can
be later converted to a string when a string is needed.
This will eventually replace the setError(int,string) used in file
engines so that strings are only created on demand.

The scope of an error is used to distinguish error codes from the C standard
library vs error codes from native calls. On some OS (e.g.windows) these
error ranges overlap.

Reviewed-By: joao