Avoid in-place convertion of images with multiple references
The decoding from image reader was assuming the image reader do not
keep the image internally. This is not true for the GIF plugins because
the previous image can be used to compose the current image.

This was causing crash on ARM because the 16 bits color depth causes
the image memory to be reduce by half. When the plugin was accessing
the memory, it assumes the images has not changed and is on 32 bits.

This patch disable the in-place conversion if a detach is required.
Regular conversion is the correct solution in this case, and it can
also be made faster by converting while copying.

Reviewed-by: Andreas Kling