[QCocoaView scrollWheel:] can end up recursing infinitely when a scrollWheel event is sent to two stacked MDI windows
The problem is the fact that we were propagating the event through the
class hierarchy. This means calling the super class implementation,
which does nothing but pass the event to the next guy, which is the
other window, which in turn does the same thing...
The point is, there is no need to propagate this event on the class
hierarchy. If anything, this should be propagated on the widgets
hierarchy, which is already handled by Qt. Removing the code so this
problem does not happen again.

Task-number: QTBUG-12952
Reviewed-by: Richard Moe Gustavsen