Doc: Reorganized the platform and compiler notes pages.
Errors and omissions will be fixed in later commits.

The idea here is to merge the compiler notes into the platform
notes because the platform is usually fixed and the compiler is
often determined by the platform. In addition, notes about
compilers are not always applicable to multiple platforms, so
collating notes by compiler does not save much space, and it only
makes maintenance of the notes more difficult.