Add support for Qt::WA_TranslucentBackground with OpenVG on Symbian^3
S^3 does not have a client side API to support semi-transparent EGL
surfaces so if this flag is enabled for a particular widget, then we
should return a raster surface instead of a VG surface. Raster surfaces
can still be semi-transparent but will be slower to render to.

Task-number: QTBUG-14400
Reviewed-by: Jani Hautakangas
Reviewed-by: Gareth Stockwell
Reviewed-by: Sami Merila