Small code cleanup.
We don't need to check for an engine twice. isJSBased() is not the best
way to check if an engine is correct it is better to use engine().

Implementation assumes that QSV::m_value is not empty, by using cast operator
we assert earlier. Btw. QSV::m_value should be private.