Correct errors in the qimagereader autotest
First, pgm and pbm formats need the same exception given to ppm format
as they are subtypes of the same format. (as text based image formats they
are vulnerable to line end differences)

Second, a workaround for the readFromDevice test case hanging on the symbian
emulator. The test expects to be able to send a whole image to a socket and
have this fit within the send/receive buffers of the TCP stack.
At least for the emulator's winsock backend, this hangs the autotest inside
a blocking write. On the Nokia 5800, the test passed so it's currently skipped
only for the emulator.

Reviewed-By: Jason Barron