Disable selection of Libraries on Windows 7 in getExistingDirectory().
The Libraries on Windows 7 are user-defined collection of folders, which
is only a logical representation. There are no file system entries
associated with it .
This patch makes sure that user cannot choose a library when using the
QFileDialog::getExistingDirectory(), in such cases it will show a
standard windows warning message.

Task-number: QTBUG-12655
Reviewed-by: Thierry Bastian