Fix crash in QApplication auto test
The pointer is lazily initialised by QWidget, but becomes invalid after
the application is destroyed.
In the QApplication autotest, QApplication is created and destroyed many times
in the same thread (while in normal use it is created at startup and destroyed
on exit)
As a result, the pointer which is stored in global data became stale and
caused a crash due to calling functions on a deleted object.

Reviewed-By: Jason Barron