Autotest for seeking beyond end of file and then writing
File engine refactor caused a regression in the TIFF image codec,
as QFile::seek() previously worked like lseek() / fseek() in posix.
But on symbian the native RFile::Seek api clamps to the end of the file
if you attempt to seek beyond there.
This test checks seek behaviour in the appropriate place, the QFile
auto test.

Reviewed-By: joao