Giving Qt a default app server when Avkon is removed
Avkon provides a default app server when an app is started as a server
app. This allows any app to be started as a server app.

The Qt Avkon removal bypassed this, calling the base-class function
instead, which prevented apps from starting as server app unless they
explictly created a server.

Now Qt also creates a default app server, which implements no services,
so that any app can be started as a server app, even if Avkon is not
present. This allows Symbian^4, configured with -no-s60, to boot.

Task-number: QTBUG-14457
Reviewed-by: Jason Barron