Fix crashes and hangs in qprocess autotest
The qprocess autotest launches various sub processes and checks their results,
but this lacked error handling.

Added a QVERIFY when opening files (if fopen returned NULL, the test fails
instead of crashing)
Added a 10s guard when waiting for a subprocess to create a file - previously
this loop did not terminate if the subprocess failed.

Reviewed-By: Markus Goetz