Cocoa: fix child window issues (QTBUG 13867, 14420, 13126)
The main problem behind these issues is the fact that when two
windows exists in a parent-child relationship (using
the [NSWindow addChild] API), Cocoa will automatically hide both
windows even when just hiding one of them. This turns out really bad
when the child is a tool window, because those will automatically be
hidden when the application becomes deactivated. And as such, the
parent will hide as well, tool or not.
So after a LOT of investigation, involving manually trying to level windows
rather than using the addChild API, the conclusing is that we cannot do
it; Cocoa and Qt just tries to outsmart each other. So instead, we now
say that only normal windows and dialogs can be part of a parent-child
relationship (which seems to be how Apple intended it as well). The
rest, and in particular tool windows, we just ignore.
This will differ from some other platforms, but at the same time, since
tool windows are on a level above other windows on mac from before, and
the docs specifies that this can be different from platform to platform,
we see it as acceptable.

Rev-By: prasanth
Rev-By: msorvig