QtScript/V8: Add API for manipulating scope and function contexts
Needed for the implementation of QScriptEngine::{push,pop}Context(),
QScriptContext::{push,pop}Scope() and friends.

This API is experimental and added only for the purpose of our
research.

The upstream v8 API only provides access to global context objects.
Internally, there are two other types of context objects:
FunctionContexts, used to implement closures, and WithContexts,
used to implement JS "with" statements. The API added in this
change provides a way to construct and access such context objects
from C++.