QS60Style: ComboBox disappears after orientation switch
This is due to that QS60Style uses desktop area for a bounding rect
of combobox content to disallow it to grow outside of screen size.
Unfortunately with some recent changes to screensize handling in
desktop widget size makes the desktop resize to happen later than
when style is asked for combobox size. So, style returns invalid
QSize for combobox content size. As a fix, if the returned size
is invalid, ask it directly from QCommonStyle.

Task-number: QTBUG-14423
Reviewed-by: Janne Anttila