Prepend epocroot to extra target tools if they start with /epoc32/
This makes it possible to define extra targets and extra compilers
commands that use tools under EPOCROOT without having to worry
whether EPOCROOT contains drive letter or not, which is demanded
by sbsv2 toolchain.

For example, if command to be used is
"%EPOCROOT%epoc32/tools/foobar.exe", the EPOCROOT should be omitted
in .pro file and the command be given simply as
"/epoc32/tools/foobar.exe".

Reviewed-by: Janne Koskinen