v8: Bad representation of SMI value in JIT code.
External::Wrap ed value are most of the case 'fake' SMI. Fake because
they have the SMI flag, but are on 64bit rather than 31 bit.
Still, they are Handle<Object>.

When saving Handle with MacroAssembler::Move, we do not want to convert
the SMI to integer. Because we would loose the other bit information.
And also because the code reading the value will expect a Handle<Object>,
so a SMI is wrong.

One typical case where this happen is in GenerateFastApiCall in
stub-cache-x64.cc, where the call_data handle is saved.
that call_data is a handle comming from v8::External::Wrap

Test that reproduces it is tst_QScriptExtQObject::findChildren
When calling findChildren several times, the Data() value would
become null at some point (if the JIT code become optimized)