Make sure QScriptPrograms are evaluated in correct context
Currently v8 doesn't provide a way to compile a script in a
way that it will pick up the current (function/scope) context
whenever the script is subsequently executed.
Disable the caching of the compiled script for now, i.e.
always evaluate the script from source, so that the behavior
is correct. (QML depends on this behavior, for example.)