Cocoa: cannot use staysOnTop flag for native file dialogs
We cannot mix staysOnTop and native file dialogs, since Cocoa will
anyway set it back to NSModalPanelWindowLevel when running it
app modal. There are ways to work around this issue, but the file
dialog also has a button for showing a "create directory" modal
panel, and this we cannot control.

Reviewed-by: cduclos