Implement seek beyond end of file for symbian
On Symbian, the Seek() function clamps to the end of file position if a
position beyond the end of file is given. Also the write functions which
take a position silently append at the end of file position if asked to
write to a start position beyond EOF.

To avoid this behaviour, we don't use seek but rather maintain our own
cursor position. When writing, we check if the cursor is beyond EOF and if
so extend the file using SetSize() before writing.

Reviewed-By: joao