Remove widget subtree from backing store tracker when reparented
When a native widget is reparented such that the value of
maybeBackingStore() changes, it and all of its native descendents
must be removed from the old QWidgetBackingStoreTracker.

If this is not done, the backing store may not be deleted when the
top-level widget is hidden, thereby consuming memory unnecessarily.

Task-number: MOBILITY-1315
Reviewed-by: Jason Barron
Reviewed-by: bnilsen