Don't allow posted events to starve native messages.
This reverts commit 9215506e6a057d8aef7415c2921214e1ba1c760d, which
caused the above behavior. The intention of this commit was to prevent
native messages from starving posted events, but as shown by the
example, it is possible for posted events to then starve native messages
(which results in very bad behavior on Windows if the user is moving or
resizing a window).

Unfortunately, my attempts to construct an autotest for this have
failed, as I've not found a way to generate the right kinds of native
messages from code.

Task-number: QTBUG-14021
Reviewed-by: prasanth