Fix QFileDialog regression in symbian
Symbian now follows the same convention as windows, where "c:" means the
working directory on the C: drive. As a result, the path "c:/" needs to be
passed to QFileInfoGatherer when getting info for the root directory.

This code already existed but only enabled for windows - now enabled for
symbian as well.

Reviewed-By: Markus Goetz