Using TLS to store QThreadData on Symbian
TLS access on Symbian is significantly faster than using
pthread_getspecific. In the case of QThreadData, the data associated
with a thread, it makes sense to store and access the pointer in TLS
rather than using the pthread thread specific functions.

However, the pthread thread specific function have some destructor
functionality that Symbian TLS does not, so Symbian still uses the
pthread functions for setting the QThreadData to get the destruction
behaviour.

This change has wide ranging performance benefits for QML execution.
With improvements shown on a number of the declarative benchmarks.
Frame drawing times on test apps are also showing some improvement.

The qthread* autotests have been run on Symbian^3 and Ubuntu 10.10

Task-number: QT-4089
Reviewed-by: Shane Kearns