Symbian shared memory could not be distinguished by numeric key
Fixes a problem where QSharedMemory objects on Symbian could not be
distinguished by numeric keys, or keys that differed in any non-alpha
way.

This is fixed by using the same native key generation system as on other
platforms, where a SHA1 hash of the original key is appended to a
version of the key with all non-alpha characters stripped.

Task-number: QT-3712
Reviewed-by: Shane Kearns