Re-enable emulated QFile::map on symbian
Memory mapping of RFile handles is not supported in current symbian OS
versions. However the "open C" libraries provide an emulated mmap()
implementation which was used by QFile::map in Qt 4.6 release.
To avoid breaking applications which rely on this function, QFile::map
will now open the file with open C as well in order to use that handle
to call mmap().
When symbian implements a file mapping API, we can switch the implementation
to use that for RFile handles.

Reviewed-By: joao