QScriptValue::construct(): Don't crash if function throws non-Object
If an exception occurs, we should ignore the result of
JSC::construct() and return the exception value, even if
the exception is not an object.
This makes the behavior match the documentation: "Calling
construct() can cause an exception to occur in the script
engine; in that case, construct() returns the value that
was thrown".

Task-number: QTBUG-14801
Reviewed-by: Jedrzej Nowacki