Avoid being killed by graphics out-of-memory monitor.
When an application requests more graphics memory, the GOOM monitor
will start looking for other applications to kill in order to recover
the requested graphics memory. The first step in this process is to
send a custom event to the application indicating that it is a
candidate for termination. If the application does not acknowledge this
event, it will be killed immediately. However, since Qt releases its
graphics memory when the app is obscured, we should indicate that we
handle this event such that Qt has a chance to finish freeing
resources. If after a certain time, the app has not freed its resources
it will be killed anyway. This essentially buys Qt some time to
avoid the case where Qt apps are killed while they are cleaning up.

Task: QTBUG-14858
Reviewed-by: Jani Hautakangas