Support tactile feeedback from QS60Style for QWidgets
Currently tactile feedback is not given for QWidgets running on Symbian
even if the native side supports this (Sym^3 and 5th Edition devices).

This task adds support for QWidgets having QS60Style. The tactile
feedback has been implemented as a plugin that the style loads when
instantiating itself for touch devices. NOTE that the feedback
is NOT supported by the emulated style, nor is the plugin interface
public, so it cannot be used outside of style.

The implementation is simplistic, since we only want to provide
stop-gap solution until 4.8 when real Qt feedback implementation
is ready. The implementation will only give feedback for touch-down
events for visible, interactive and enabled widgets. Sliders and
scrollbars will use sensitive feedback (slightly less aggressive)
and all others will use basic feedback. Note that Sym^3 adds tens
of different feedback categories, which this plugin ignores
as we want to share the same implementation for Sym^3 and 5th ed.

In distributed Qt package there is no tactile feedback for 5th edition,
due to package creation limitations. Support can be added manually by
re-compiling Qt on top of 5th Edition SDK.

Task-number: QT-4037
Reviewed-by: Jani Hautakangas
Reviewed-by: Janne Koskinen