Prevented threading related crash in OpenGL module.
If the last shallow copy of a QImage which is cached in the
QGLTextureCache is destroyed in a thread at the same time as the
QGLContext which the texture was initialized in is active in a different
thread, the QImage thread incorrectly tries to make the context active
in two threads at once. To prevent this from happening it's best to
always do the texture clean-up in the main thread.

Task-number: QT-4238
Reviewed-by: Eskil Abrahamsen Blomfeldt