QNAM: Do not cache network connections per username.
If user information is present in the Url, QNetworkAccessHttpBackend
will cache the connection with a key in the format of user@hostname.

This causes the following tests in QtWebKit to fail:

LayoutTests/http/tests/xmlhttprequest/re-login.html
LayoutTests/http/tests/xmlhttprequest/re-login-async.html

The tests call the same url with user1 credentials, no credentials,
user 2 credentials and no credentials. In the second and fourth calls
Qt should use the most recently stored credentials. For the fourth
call it currently uses user1's credentials instead of user2's. This
is because it retrieves the cached connection from the third call
and uses the user/pass in that connection's authenticator - which is
user1's.

In order to retrieve the connection with the most recently used
credentials userinfo should be removed when caching the connection
in QNetworkAccessCache. This will allow Qt to retrieve the connection
used in the third call, in fact the same cached connection for all
four calls.

Merge-request: 623
Reviewed-by: Markus Goetz <Markus.Goetz@nokia.com>