Fixed event starvation with pointer events on Symbian.
The CEikonEnv active object is used by Symbian to indicate that new
window server events are available, in this case pointer events.
Since this object was being controlled by Symbian, it was not subject
to Qt's round robin queue. In a case where the event handler for the
pointer event would take very long, it was possible for the active
object to be completed again before the previous handler was
finished, causing it to be called in a loop.

This was fixed by making a new Qt-owned class as a subclass of
CEikonEnv, and in its RunL, make sure that the object follows the
round-robin queue.

Because the round robin code is located in a subclass of CEikonEnv
that we own, the fix will have no effect if Qt is loaded after the
Eikon framework (such as a Qt plugin in a non-Qt app).

Task:     QT-4077
RevBy:    mread
AutoTest: N/A, system events