Qt/Symbian: QInputDialog layout and size are wrong
The QInputDialog with anything but QLineEdit as the input widget,
will show up as minimal size on device/emulator. Also, the dialog
looks bad since the widgets are deployed very close to the dialog
borders.

As a fix, layout will not get the size policy set on Symbian, but
the dialog itself is stretched to occupy screen width and thus, layout
grows to occupy the available size. Additionally, style will set layout
pixel metrics for dialogs larger than for other widgets, to avoid
the layout from putting widgets very close to dialog border.

Task-number: QTBUG-14331
Reviewed-by: Janne Anttila