Prevent access to non-existent memory in QGL2PEXVertexArray
QGL2PEXVertexArray::addClosingLine() uses using QDataBuffer::add() to
add a QPointF at the end, with a const reference to another value in
the same array. As add() resizes and possibly relocates the buffer, an
already-released memory location may be accessed.

This change copies the value into a temporary variable before resizing
the array.

Task-number: QTBUG-14944
Reviewed-by: Samuel