Fix bidi PDF mark support in Core Text shaper
When Core Text shaping a string like LRE ... PDF, it won't leave
a zero width glyph for the PDF mark, it is not consistent with
HarfBuzz and will cause issues in our text engine (the last glyph
of the string will be considered as a ligature). This hack
appended such a zero width glyph to fix that.

Reviewed-by: Eskil