Make QThreadStorage supports value type and not only pointers.
Handling value type is much more natural than handling pointer.
It was not possible to use normal type in QThreadStorage previously
(probably because some compiler would not support it?)

This should ease a lot the use of QThreadStorage and make it more
intuitive.

The only problem was the QThreadStorage::deleteData() that would
not compile for nonn-pointer. This is now fixed

Also updated the documentation.

Reviewed-by: Joao
Reviewed-by: Brad
Task-number: QTBUG-15033